feat: 게임 중간에도 status 명령어로 점수를 확인할 수 있도록 변경
