+++ src/main/java/chess/domain/ChessGame.java
@@ -49,13 +49,13 @@ public class ChessGame {
 
     private void executeCommand(Board board) {
         List<String> input = List.of(InputView.requestCommand().split(" "));
-
         if (Command.inGameCommand(input.get(0)) == Command.END) {
             board.terminate();
-            return;
         }
-
-        if (input.size() == 3) {
+        if (Command.inGameCommand(input.get(0)) == Command.STATUS) {
+            OutputView.printScore(board);
+        }
+        if (Command.inGameCommand(input.get(0)) == Command.MOVE && input.size() == 3) {
             board.move(new Position(input.get(1)), new Position(input.get(2)));
         }
     }
